# docs/swagger.yaml
definitions:
  server.AnswerRequest:
    properties:
      answer_value:
        enum:
        - "Yes"
        - "No"
        - Pass
        example: "Yes"
        type: string
    required:
    - answer_value
    type: object
...
info:
  contact: {}
paths:
  /api/question/{id}:
    post:
      consumes:
      - application/json
      description: Create an answer for a specific question in a questionnaire
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      - description: Answer data
        in: body
        name: answer
        required: true
        schema:
          $ref: '#/definitions/server.AnswerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Answer created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Question not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Answer a question
      tags:
      - questions
...
swagger: "2.0"